component folder tree refactor
